Re-add timestamp definition
A few months ago the timestamp definition was moved to a separate file. Here I've added it back in because our system is not able to reach out to the web to use the schema. I'm not sure if the validation library I'm using is able to load and validate against multiple schema files at once, but thought I'd propose this as this is what we're currently going forward with.